Add files via upload
The "full_test_sys.py" file consolidates elements of the other three scripts to create a single system with which to test an already trained model. The script uses two files: "face-embeddings.npz" and "face classifier.sav." Both of which can be generated by running the previous three scripts on the full face database. 
Then, to test classification on any new images, simply drop said images into a folder entitled "test_images" and run the script in the same directory.